Moved code from submitted_content_controller to submitted_files_helper
